--- cub3D_working/.gitignore
@@ -1,18 +1,0 @@
-# binaries
-cub3D
-libft/libft.a
-# objects
-objs/*/*.o
-objs/*/*.d
-objs/*/*.cc
-libft/objs/*.o
-# cache
-.cache/
-.cache/*
-# debug files
-gdb.txt
-# compile_commands
-compile_commands.json
-# submodules
-minilibx-linux/
-libft/
