--- cub3D_working/includes/args.h
@@ -1,18 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   args.h                                             :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: dpentlan <dpentlan@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/11/17 18:41:01 by dpentlan          #+#    #+#             */
-/*   Updated: 2023/11/17 18:42:22 by dpentlan         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#ifndef ARGS_H
-# define ARGS_H
-
-void	ft_arg_parse(int argc, char **argv);
-
-#endif
