--- cub3D_working/includes/map_parsing.h
@@ -1,18 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   map_parsing.h                                      :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: dpentlan <dpentlan@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/11/17 18:40:07 by dpentlan          #+#    #+#             */
-/*   Updated: 2023/11/17 18:40:46 by dpentlan         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#ifndef MAP_PARSING_H
-# define MAP_PARSING_H
-
-void	ft_init_map(void);
-
-#endif
