--- cub3D_working/srcs/map_parsing/map_parsing.c
@@ -1,18 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   map_parsing.c                                      :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: dpentlan <dpentlan@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/11/17 18:36:30 by dpentlan          #+#    #+#             */
-/*   Updated: 2023/11/17 18:37:24 by dpentlan         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "cub3D.h"
-
-void	ft_init_map(void)
-{
-	return ;
-}
