+++ srcs/args/arg_parsing.c
@@ -1,0 +1,20 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   arg_parsing.c                                      :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: dpentlan <dpentlan@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/11/17 18:37:54 by dpentlan          #+#    #+#             */
+/*   Updated: 2023/11/17 18:42:09 by dpentlan         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "cub3D.h"
+
+void	ft_arg_parse(int argc, char **argv)
+{
+	(void) argc;
+	(void) argv;
+	return ;
+}
